+++ NEWS.md
@@ -1,3 +1,8 @@
+# heatwaveR 0.4.5.9000 (2021-01-31)
+
+-   Added a `categories` argument to `detect_event()`
+    -   This allows one to determine the categories of events directly, rather than as a second step using the `category()` function
+
 # heatwaveR 0.4.5 (2021-01-23)
 
 -   Created a new vignette that shows how to save gridded MHW results to a NetCDF file